Move backfill function to lib folder for reuse
